iot_web_dashboard/requirements.txt
Flask
Flask-CORS
boto3
python-dotenv
paho-mqtt